Oluşturuldu.

+++ src/NotOrtalamasi.java
@@ -20,7 +20,7 @@ public class NotOrtalamasi {
         muzik = inp.nextInt();
 
         // Ortalamayı hesaplayıp sınıf geçme durumuyla beraber çıkarttık.
-        int ort = (mat+fizik+kimya+turkce+tarih+muzik)/6;
+        double ort = (mat+fizik+kimya+turkce+tarih+muzik)/6.0;
         System.out.println("Ortalamanız : "+ort);
         System.out.println(ort>=60 ? "Sınıfı Geçtiniz!" : "Sınıfta Kaldınız!");
